+++ pkg/store/tsdb_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/fortytw2/leaktest"
 	"github.com/prometheus/prometheus/pkg/labels"
+	"github.com/prometheus/prometheus/pkg/timestamp"
 	"github.com/thanos-io/thanos/pkg/component"
 	"github.com/thanos-io/thanos/pkg/store/storepb"
 	"github.com/thanos-io/thanos/pkg/testutil"
@@ -188,40 +189,93 @@ func TestTSDBStore_LabelNames(t *testing.T) {
 	testutil.Ok(t, err)
 
 	appender := db.Appender()
-	addLabels := func(lbs []string) {
+	addLabels := func(lbs []string, timestamp int64) {
 		if len(lbs) > 0 {
-			_, err = appender.Add(labels.FromStrings(lbs...), math.MaxInt64, 1)
+			_, err = appender.Add(labels.FromStrings(lbs...), timestamp, 1)
 			testutil.Ok(t, err)
 		}
 	}
 
 	tsdbStore := NewTSDBStore(nil, nil, db, component.Rule, labels.FromStrings("region", "eu-west"))
 
+	now := time.Now()
+	head := db.Head()
 	for _, tc := range []struct {
 		title         string
 		labels        []string
 		expectedNames []string
+		timestamp     int64
+		start         func() int64
+		end           func() int64
 	}{
 		{
 			title:         "no label in tsdb",
 			labels:        []string{},
 			expectedNames: []string{},
+			timestamp:     now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
 		},
 		{
 			title:         "add one label",
 			labels:        []string{"foo", "foo"},
 			expectedNames: []string{"foo"},
+			timestamp:     now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
 		},
 		{
 			title:  "add another label",
 			labels: []string{"bar", "bar"},
 			// We will get two labels here.
 			expectedNames: []string{"bar", "foo"},
+			timestamp:     now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
+		},
+		{
+			title:         "query range outside tsdb head",
+			labels:        []string{},
+			expectedNames: []string{},
+			timestamp:     now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return head.MinTime() - 1
+			},
+		},
+		{
+			title:         "get all labels",
+			labels:        []string{"buz", "buz"},
+			expectedNames: []string{"bar", "buz", "foo"},
+			timestamp:     now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
 		},
 	} {
 		if ok := t.Run(tc.title, func(t *testing.T) {
-			addLabels(tc.labels)
-			resp, err := tsdbStore.LabelNames(ctx, &storepb.LabelNamesRequest{})
+			addLabels(tc.labels, tc.timestamp)
+			resp, err := tsdbStore.LabelNames(ctx, &storepb.LabelNamesRequest{
+				Start: tc.start(),
+				End:   tc.end(),
+			})
 			testutil.Ok(t, err)
 			testutil.Equals(t, tc.expectedNames, resp.Names)
 			testutil.Equals(t, 0, len(resp.Warnings))
@@ -243,43 +297,86 @@ func TestTSDBStore_LabelValues(t *testing.T) {
 	testutil.Ok(t, err)
 
 	appender := db.Appender()
-	addLabels := func(lbs []string) {
+	addLabels := func(lbs []string, timestamp int64) {
 		if len(lbs) > 0 {
-			_, err = appender.Add(labels.FromStrings(lbs...), math.MaxInt64, 1)
+			_, err = appender.Add(labels.FromStrings(lbs...), timestamp, 1)
 			testutil.Ok(t, err)
 		}
 	}
 
 	tsdbStore := NewTSDBStore(nil, nil, db, component.Rule, labels.FromStrings("region", "eu-west"))
 
+	now := time.Now()
+	head := db.Head()
 	for _, tc := range []struct {
 		title          string
 		addedLabels    []string
 		queryLabel     string
 		expectedValues []string
+		timestamp      int64
+		start          func() int64
+		end            func() int64
 	}{
 		{
 			title:          "no label in tsdb",
 			addedLabels:    []string{},
 			queryLabel:     "foo",
 			expectedValues: []string{},
+			timestamp:      now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
 		},
 		{
 			title:          "add one label value",
 			addedLabels:    []string{"foo", "test"},
 			queryLabel:     "foo",
 			expectedValues: []string{"test"},
+			timestamp:      now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
 		},
 		{
 			title:          "add another label value",
 			addedLabels:    []string{"foo", "test1"},
 			queryLabel:     "foo",
 			expectedValues: []string{"test", "test1"},
+			timestamp:      now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return timestamp.FromTime(maxTime)
+			},
+		},
+		{
+			title:          "query time range outside head",
+			addedLabels:    []string{},
+			queryLabel:     "foo",
+			expectedValues: []string{},
+			timestamp:      now.Unix(),
+			start: func() int64 {
+				return timestamp.FromTime(minTime)
+			},
+			end: func() int64 {
+				return head.MinTime() - 1
+			},
 		},
 	} {
 		if ok := t.Run(tc.title, func(t *testing.T) {
-			addLabels(tc.addedLabels)
-			resp, err := tsdbStore.LabelValues(ctx, &storepb.LabelValuesRequest{Label: tc.queryLabel})
+			addLabels(tc.addedLabels, tc.timestamp)
+			resp, err := tsdbStore.LabelValues(ctx, &storepb.LabelValuesRequest{
+				Label: tc.queryLabel,
+				Start: tc.start(),
+				End:   tc.end(),
+			})
 			testutil.Ok(t, err)
 			testutil.Equals(t, tc.expectedValues, resp.Values)
 			testutil.Equals(t, 0, len(resp.Warnings))